udpated folder structre for fastapi exercise
