+++ .pre-commit-config.yaml
@@ -1,0 +1,31 @@
+repos:
+  - repo: https://github.com/pre-commit/pre-commit-hooks
+    rev: v3.4.0
+    hooks:
+      - id: trailing-whitespace
+      - id: end-of-file-fixer
+      - id: check-yaml
+      - id: check-json
+      - id: check-added-large-files
+
+  # Add more repositories and hooks as needed
+  # For example, if you're using black for formatting and flake8 for linting:
+  # - repo: https://github.com/psf/black
+  #   rev: 21.9b0
+  #   hooks:
+  #     - id: black
+  #       language_version: python3
+  # - repo: https://github.com/PyCQA/flake8
+  #   rev: 4.0.1
+  #   hooks:
+  #     - id: flake8
+
+  # Example for isort if you're using Python
+  # - repo: https://github.com/pycqa/isort
+  #   rev: 5.8.0
+  #   hooks:
+  #     - id: isort
+  #       additional_dependencies: ["isort==5.8.0"]
+      #   # See https://pre-commit.com/hooks.html#additional-dependencies
+
+  # Add more hooks as needed
