+++ constants/keys.py
@@ -1,0 +1,2 @@
+SECRET_KEY = "df9ad6883d0f1bcbbd99c2dbfc916bbaf17a7454b27510f3497d4217f924b3d7"
+ALGORITHM = "HS256"
