+++ database/session.py
@@ -1,0 +1,28 @@
+import os
+from dotenv import load_dotenv
+from sqlalchemy import create_engine
+from sqlalchemy.orm import sessionmaker
+
+load_dotenv()
+
+# Constructing SQLite database URL
+URL_DATABASE = "sqlite:///./sqlite_db.db"
+# URL_DATABASE = f"postgresql://{os.getenv('DB_USER')}:{os.getenv('DB_PASSWORD')}@{os.getenv('DB_HOST')}:{os.getenv('DB_PORT')}/{os.getenv('DB_NAME')}"
+
+engine = create_engine(URL_DATABASE)
+
+SessionLocal = sessionmaker(autocommit=False, autoflush=False, bind=engine)
+
+
+def get_db():
+    """
+    Function to yield a database session.
+
+    Yields:
+        Session: SQLAlchemy database session.
+    """
+    db = SessionLocal()
+    try:
+        yield db
+    finally:
+        db.close()
