+++ models/base.py
@@ -1,0 +1,22 @@
+from sqlalchemy.ext.declarative import declarative_base
+from sqlalchemy import Column, Integer, DateTime, ForeignKey
+from sqlalchemy.orm import relationship
+from datetime import datetime
+
+Base = declarative_base()
+
+
+class TimestampMixin:
+    created_at = Column(DateTime, default=datetime.utcnow)
+    updated_at = Column(DateTime, default=datetime.utcnow, onupdate=datetime.utcnow)
+
+
+class UserTrackingMixin:
+    created_by_id = Column(Integer, ForeignKey('users.id'))
+    updated_by_id = Column(Integer, ForeignKey('users.id'))
+    created_by = relationship("User", foreign_keys=[created_by_id])
+    updated_by = relationship("User", foreign_keys=[updated_by_id])
+
+
+class BaseModel(Base, TimestampMixin, UserTrackingMixin):
+    __abstract__ = True
