+++ routers/logger.py
@@ -1,0 +1,24 @@
+import logging
+
+# Create logger
+logger = logging.getLogger(__name__)
+
+# Configure logger
+logger.setLevel(logging.INFO)
+
+# Define a formatter
+formatter = logging.Formatter('%(asctime)s - %(name)s - %(levelname)s - %(message)s')
+
+# Create a file handler and set the formatter
+file_handler = logging.FileHandler('app.log')
+file_handler.setLevel(logging.INFO)
+file_handler.setFormatter(formatter)
+
+# Add the file handler to the logger
+logger.addHandler(file_handler)
+
+# Optionally, add a console handler for logging to console
+console_handler = logging.StreamHandler()
+console_handler.setLevel(logging.INFO)
+console_handler.setFormatter(formatter)
+logger.addHandler(console_handler)
