+++ schemas/user_role.py
@@ -1,0 +1,8 @@
+from pydantic import BaseModel
+
+
+class UserRoleCreate(BaseModel):
+    """
+    Model for creating a new UserRole.
+    """
+    role_name: str
