+++ schemas/user_technology.py
@@ -1,0 +1,8 @@
+from pydantic import BaseModel
+
+
+class UserTechnologyCreate(BaseModel):
+    """
+    Model for creating a new UserTechnology.
+    """
+    technology_name: str
