+++ utils/hash_pwd.py
@@ -1,0 +1,7 @@
+import bcrypt
+
+
+def hash_password(password: str) -> str:
+    salt = bcrypt.gensalt()
+    hashed_password = bcrypt.hashpw(password.encode('utf-8'), salt)
+    return hashed_password.decode('utf-8')
