+++ utils/verify_pwd.py
@@ -1,0 +1,5 @@
+import bcrypt
+
+
+def verify_password(plain_password: str, hashed_password: str) -> bool:
+    return bcrypt.checkpw(plain_password.encode('utf-8'), hashed_password.encode('utf-8'))
